pyproject.toml
[project]
name = "one-time-secrets"
version = "0.1.0"
description = ""
authors = [
    {name = "ratmir",email = "ratmir2020.1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "cryptography (>=44.0.2,<45.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^1.1.0"
mypy = "^1.15.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
